k8sautodeploy/requirements.txt
jinja2~=3.1.1
PyYAML~=6.0